Cambios
Se realizan cambios de validaciones y ajuste codigo

+++ src/cl/billeteraVirtual/clases/Conversor.java
@@ -2,6 +2,6 @@ package cl.billeteraVirtual.clases;
 
 public interface Conversor {
 
-    void convertirSaldo(int saldo);
+    void convertirSaldo(double saldo);
 
 }
