+++ src/cl/billeteraVirtual/clases/Transaccion.java
@@ -1,9 +1,9 @@
 package cl.billeteraVirtual.clases;
 
 public interface Transaccion {
-    void ingresarSaldo(int monto);
+    void ingresarSaldo(double monto);
 
-    void retirarSaldo(int monto);
+    void retirarSaldo(double monto);
 
     void mostrarSaldo();
